+++ src/main.rs
@@ -5,17 +5,20 @@ use std::io::IsTerminal;
 
 use clap::Parser;
 use clap::Subcommand;
+use clap_verbosity_flag::Verbosity;
+use colored::Colorize;
 use git_testament::git_testament;
 use git_testament::render_testament;
-
-pub mod commands;
+use sprocket::commands;
+use tracing_log::AsTrace;
 
 git_testament!(TESTAMENT);
 
 #[derive(Subcommand)]
 #[allow(clippy::large_enum_variant)]
 enum Commands {
-    /// Checks the syntactic validity of Workflow Description Language files.
+    /// Checks a WDL file (or a directory containing WDL files) and reports
+    /// diagnostics.
     Check(commands::check::CheckArgs),
 
     /// Lints Workflow Description Language files.
@@ -34,39 +37,25 @@ struct Cli {
     #[command(subcommand)]
     pub command: Commands,
 
-    /// Only errors are printed to the stderr stream.
-    #[arg(short, long, global = true)]
-    pub quiet: bool,
-
-    /// All available information, including debug information, is printed to
-    /// stderr.
-    #[arg(short, long, global = true)]
-    pub verbose: bool,
+    #[command(flatten)]
+    verbose: Verbosity,
 }
 
 pub async fn inner() -> anyhow::Result<()> {
     let cli = Cli::parse();
 
-    let level = if cli.verbose {
-        tracing::Level::DEBUG
-    } else if cli.quiet {
-        tracing::Level::ERROR
-    } else {
-        tracing::Level::INFO
-    };
-
     tracing_log::LogTracer::init()?;
 
     let subscriber = tracing_subscriber::fmt::Subscriber::builder()
-        .with_max_level(level)
+        .with_max_level(cli.verbose.log_level_filter().as_trace())
         .with_writer(std::io::stderr)
         .with_ansi(stderr().is_terminal())
         .finish();
     tracing::subscriber::set_global_default(subscriber)?;
 
     match cli.command {
-        Commands::Check(args) => commands::check::check(args),
-        Commands::Lint(args) => commands::check::lint(args),
+        Commands::Check(args) => commands::check::check(args).await,
+        Commands::Lint(args) => commands::check::lint(args).await,
         Commands::Explain(args) => commands::explain::explain(args),
         Commands::Analyzer(args) => commands::analyzer::analyzer(args).await,
     }
@@ -74,7 +63,15 @@ pub async fn inner() -> anyhow::Result<()> {
 
 #[tokio::main]
 pub async fn main() {
-    if let Err(err) = inner().await {
-        eprintln!("error: {}", err);
+    if let Err(e) = inner().await {
+        eprintln!(
+            "{error}: {e:?}",
+            error = if std::io::stderr().is_terminal() {
+                "error".red().bold()
+            } else {
+                "error".normal()
+            }
+        );
+        std::process::exit(1);
     }
 }